make separate common header that defines macros
