+++ python/bvhar/_src/_ols.cpp
@@ -1,54 +1,11 @@
 #include <pybind11/eigen.h>
 #include <ols.h>
 
-py::dict bvhar::MultiOls::returnOlsRes() {
-	this->fit();
-	return py::dict(
-		py::arg("coefficients") = this->coef,
-		py::arg("fitted.values") = this->yhat,
-		py::arg("residuals") = this->resid,
-		py::arg("covmat") = this->cov,
-		py::arg("df") = this->dim_design,
-		py::arg("m") = this->dim,
-		py::arg("obs") = this->num_design,
-		py::arg("y0") = this->response
-	);
-}
-
-py::dict bvhar::OlsVar::returnOlsRes() {
-	py::dict ols_res = this->_ols->returnOlsRes();
-	ols_res["p"] = this->lag;
-	ols_res["totobs"] = this->data.rows();
-	ols_res["process"] = "VAR";
-	ols_res["type"] = this->const_term ? "const" : "none";
-	ols_res["design"] = this->design;
-	ols_res["y"] = this->data;
-	return ols_res;
-}
-
-py::dict bvhar::OlsVhar::returnOlsRes() {
-	py::dict ols_res = this->_ols->returnOlsRes();
-	ols_res["p"] = 3;
-	ols_res["week"] = this->week;
-	ols_res["month"] = this->month;
-	ols_res["totobs"] = this->data.rows();
-	ols_res["process"] = "VHAR";
-	ols_res["type"] = this->const_term ? "const" : "none";
-	ols_res["HARtrans"] = this->har_trans;
-	ols_res["design"] = this->var_design;
-	ols_res["y"] = this->data;
-	return ols_res;
-}
-
 PYBIND11_MODULE(_ols, m) {
 	m.doc() = "OLS for VAR and VHAR";
 
   py::class_<bvhar::MultiOls>(m, "MultiOls")
     .def(py::init<const Eigen::MatrixXd&, const Eigen::MatrixXd&>())
-		// .def("estimateCoef", &bvhar::MultiOls::estimateCoef)
-		// .def("fitObs", &bvhar::MultiOls::fitObs)
-		// .def("estimateCov", &bvhar::MultiOls::estimateCov)
-		// .def("fit", &bvhar::MultiOls::fit)
     .def("returnOlsRes", &bvhar::MultiOls::returnOlsRes);
 	
 	py::class_<bvhar::LltOls, bvhar::MultiOls>(m, "LltOls")
